refactor: move `TestCase` to `tempest/generation`
