+++ phpstan.neon.dist
@@ -22,6 +22,9 @@ parameters:
 			message: '#.*[Rr]eadonly.*#'
 			path: src/Tempest/Core/src/Composer.php
 		-
+			message: '#.*final#'
+			path:  src/Tempest/Generation/tests/TestCase.php
+		-
 			message: '#.*#'
 			path: src/Tempest/Http/Exceptions/HttpExceptionHandler.php
 		-
