+++ src/Tempest/Generation/tests/ClassGeneratorTest.php
@@ -9,7 +9,6 @@ use Tempest\Generation\ClassManipulator;
 use Tempest\Generation\Tests\Fixtures\CreateMigrationsTable;
 use Tempest\Generation\Tests\Fixtures\Database\MigrationModel;
 use Tempest\Generation\Tests\Fixtures\TestAttribute;
-use Tests\Tempest\TestCase;
 
 /**
  * @internal
